Keep runtime test suites within size guard
